+++ system/dev/bus/platform/platform-proxy.h
@@ -39,7 +39,7 @@ public:
                     platform_proxy_rsp_t* resp, uint32_t resp_length,
                     zx_handle_t* in_handles, uint32_t in_handle_count,
                     zx_handle_t* out_handles, uint32_t out_handle_count,
-                    uint32_t* out_actual);
+                    size_t* out_actual);
 
     inline zx_status_t Rpc(uint32_t device_id, platform_proxy_req_t* req, uint32_t req_length,
                            platform_proxy_rsp_t* resp, uint32_t resp_length) {
